# shared/schemas/api.openapi.yaml
openapi: 3.0.3
info:
  title: QuickSight Assets Portal API
  description: API for managing and viewing QuickSight assets
  version: 1.0.0
servers:
  - url: https://api.quicksight-portal.example.com
    description: Production API
  - url: http://localhost:3000
    description: Local development

# Global security - applies to all endpoints unless overridden
security:
  - BearerAuth: []
paths:
...
  /api/assets:
    get:
      summary: List all assets
      parameters:
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/AssetType'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AssetStatus'
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssetListItem'
...
  /api/groups/{groupName}/assets:
    get:
      summary: Get assets accessible by a group
      description: Returns all assets that a group has access to through direct permissions or folder membership
      parameters:
        - name: groupName
          in: path
          required: true
          schema:
            type: string
          description: The name of the group
        - name: assetType
          in: query
          schema:
            $ref: '#/components/schemas/AssetType'
          description: Filter by asset type (optional)
      responses:
        '200':
          description: List of accessible assets
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    required: [groupName, totalAssets, assets]
                    properties:
                      groupName:
                        type: string
                      totalAssets:
                        type: integer
                      assetsByType:
                        type: object
                        properties:
                          dashboards:
                            type: integer
                          datasets:
                            type: integer
                          analyses:
                            type: integer
                          datasources:
                            type: integer
                          folders:
                            type: integer
                      assets:
                        type: array
                        items:
                          type: object
                          required: [assetId, assetType, assetName, accessType]
                          properties:
                            assetId:
                              type: string
                            assetType:
                              $ref: '#/components/schemas/AssetType'
                            assetName:
                              type: string
                            arn:
                              type: string
                            accessType:
                              type: string
                              enum: [direct, folder_inherited]
                              description: How the group has access to this asset
                            folderPath:
                              type: string
                              description: Folder path if access is inherited from folder
                            permissions:
                              type: array
                              items:
                                type: string
                              description: List of permissions the group has on this asset
        '404':
          description: Group not found
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for API authentication
  schemas:
...
    AssetType:
      type: string
      enum: [dashboard, analysis, dataset, datasource, folder, user, group]
    AssetStatus:
      type: string
      enum: [active, archived, deleted]
    EnrichmentStatus:
      type: string
      enum: [skeleton, enriched, partial, metadata-update]

    # Paginated response
...
    AssetListItem:
      type: object
      required: [id, name, type, status, createdTime, lastUpdatedTime, lastExportTime, enrichmentStatus]
      properties:
        id:
          type: string
          description: Asset unique identifier
        name:
          type: string
          description: Asset display name
        type:
          $ref: '#/components/schemas/AssetType'
        status:
          $ref: '#/components/schemas/AssetStatus'
        createdTime:
          type: string
          format: date-time
          description: When asset was created in QuickSight
        lastUpdatedTime:
          type: string
          format: date-time
          description: When asset was last updated in QuickSight (QuickSight's lastModifiedTime)
        lastExportTime:
          type: string
          format: date-time
          description: When asset was last exported to our portal (our timestamp)
        enrichmentStatus:
          $ref: '#/components/schemas/EnrichmentStatus'
        enrichmentTimestamps:
          type: object
          description: Timestamps for when each enrichment step was completed
          properties:
            definition:
              type: string
              format: date-time
              description: When asset definition was enriched
            permissions:
              type: string
              format: date-time
              description: When permissions were enriched
            tags:
              type: string
              format: date-time
              description: When tags were enriched
            lineage:
              type: string
              format: date-time
              description: When lineage was enriched
            views:
              type: string
              format: date-time
              description: When view stats were enriched
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
          default: []
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
          default: []
          description: Asset permissions

    # Archived asset response (extends AssetListItem)
...
    Tag:
      type: object
      required: [key, value]
      properties:
        key:
          type: string
          description: Tag key
        value:
          type: string
          description: Tag value
    Permission:
      type: object
      required: [principal, principalType, actions]
      properties:
        principal:
          type: string
          description: Principal ARN or identifier
        principalType:
          type: string
          enum: [USER, GROUP, NAMESPACE, PUBLIC]
          description: Type of principal
        actions:
          type: array
          items:
            type: string
          description: List of allowed actions

    # Activity types
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
  responses:
...
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...